Function console::colors_enabled_stderr

pub fn colors_enabled_stderr() -> bool

Returns true if colors should be enabled for stderr.

This honors the clicolors spec.

CLICOLOR != 0: ANSI colors are supported and should be used when the program isn’t piped.

CLICOLOR == 0: Don’t output ANSI color escape codes.

CLICOLOR_FORCE != 0: ANSI colors should be enabled no matter what.