Function console::new_serial_console

pub fn new_serial_console<S, I, O>(
    name: S,
    input_stream: I,
    output_stream: O
) -> Console<I, O, TtyBackend<O>>where
    S: Into<String>,
    I: Read,
    O: Write + Send + 'static,

Creates a new console and a new [TextTerminal] that reads input data from the given input_stream.

The terminal created by this function will use a [TtyBackend] that writes terminal output and control commands to the given output_stream.