console

Function user_attended

pub fn user_attended() -> bool

A fast way to check if the application has a user attended for stdout.

This means that stdout is connected to a terminal instead of a file or redirected by other means. This is a shortcut for checking the is_attended feature on the stdout terminal.