console

Struct Style

pub struct Style { /* private fields */ }

A stored style that can be applied.

Implementations

impl Style

pub const fn new() -> Self

Returns an empty default style.

pub fn from_dotted_str(s: &str) -> Self

Creates a style from a dotted string.

Effectively the string is split at each dot and then the terms in between are applied. For instance red.on_blue will create a string that is red on blue background. 9.on_12 is the same, but using 256 color numbers. Unknown terms are ignored.

pub fn apply_to<D>(&self, val: D) -> StyledObject<D>

Apply the style to something that can be displayed.

pub const fn force_styling(self, value: bool) -> Self

Forces styling on or off.

This overrides the automatic detection.

pub const fn for_stderr(self) -> Self

Specifies that style is applying to something being written on stderr.

pub const fn for_stdout(self) -> Self

Specifies that style is applying to something being written on stdout.

This is the default behaviour.

pub const fn fg(self, color: Color) -> Self

Sets a foreground color.

pub const fn bg(self, color: Color) -> Self

Sets a background color.

pub fn attr(self, attr: Attribute) -> Self

Adds a attr.

pub const fn black(self) -> Self

pub const fn red(self) -> Self

pub const fn green(self) -> Self

pub const fn yellow(self) -> Self

pub const fn blue(self) -> Self

pub const fn magenta(self) -> Self

pub const fn cyan(self) -> Self

pub const fn white(self) -> Self

pub const fn color256(self, color: u8) -> Self

pub const fn bright(self) -> Self

pub const fn on_black(self) -> Self

pub const fn on_red(self) -> Self

pub const fn on_green(self) -> Self

pub const fn on_yellow(self) -> Self

pub const fn on_blue(self) -> Self

pub const fn on_magenta(self) -> Self

pub const fn on_cyan(self) -> Self

pub const fn on_white(self) -> Self

pub const fn on_color256(self, color: u8) -> Self

pub const fn on_bright(self) -> Self

pub fn bold(self) -> Self

pub fn dim(self) -> Self

pub fn italic(self) -> Self

pub fn underlined(self) -> Self

pub fn blink(self) -> Self

pub fn blink_fast(self) -> Self

pub fn reverse(self) -> Self

pub fn hidden(self) -> Self

pub fn strikethrough(self) -> Self