Crate console

Creation and management of virtual consoles or terminals atop Theseus.

Structs

Console

Functions

ignore_serial_port_input

Configures the console connection listener to ignore inputs from the given serial port.

new_serial_console

Creates a new console and a new [TextTerminal] that reads input data from the given input_stream.

start_connection_detection

Starts a new task that detects new console connections by waiting for new data to be received on serial ports.