Struct console::Term

pub struct Term { /* private fields */ }

Abstraction around a terminal.

A terminal can be cloned. If a buffer is used it’s shared across all clones which means it largely acts as a handle.

Implementations

impl Term

pub fn stdout() -> Term

Return a new unbuffered terminal

pub fn stderr() -> Term

Return a new unbuffered terminal to stderr

pub fn buffered_stdout() -> Term

Return a new buffered terminal

pub fn buffered_stderr() -> Term

Return a new buffered terminal to stderr

pub fn read_write_pair<R, W>(read: R, write: W) -> Term where
R: Read + Debug + AsRawFd + Send + 'static,
W: Write + Debug + AsRawFd + Send + 'static,

Return a terminal for the given Read/Write pair styled-like Stderr.

pub fn read_write_pair_with_style<R, W>(read: R, write: W, style: Style) -> Term where
R: Read + Debug + AsRawFd + Send + 'static,
W: Write + Debug + AsRawFd + Send + 'static,

Return a terminal for the given Read/Write pair.

pub fn style(&self) -> Style

Returns the style for the term

pub fn target(&self) -> TermTarget

Returns the target

pub fn write_line(&self, s: &str) -> Result<()>

Writes a string to the terminal and adds a newline.

pub fn read_char(&self) -> Result<char>

Read a single character from the terminal

This does not echo the character and blocks until a single character is entered.

pub fn read_key(&self) -> Result<Key>

Read a single key form the terminal.

This does not echo anything. If the terminal is not user attended the return value will always be the unknown key.

pub fn read_line(&self) -> Result<String>

Read one line of input.

This does not include the trailing newline. If the terminal is not user attended the return value will always be an empty string.

pub fn read_line_initial_text(&self, initial: &str) -> Result<String>

Read one line of input with initial text.

This does not include the trailing newline. If the terminal is not user attended the return value will always be an empty string.

pub fn read_secure_line(&self) -> Result<String>

Read securely a line of input.

This is similar to read_line but will not echo the output. This also switches the terminal into a different mode where not all characters might be accepted.

pub fn flush(&self) -> Result<()>

Flushes internal buffers.

This forces the contents of the internal buffer to be written to the terminal. This is unnecessary for unbuffered terminals which will automatically flush.

pub fn is_term(&self) -> bool

Checks if the terminal is indeed a terminal.

pub fn features(&self) -> TermFeatures<'_>

Checks for common terminal features.

pub fn size(&self) -> (u16, u16)

Returns the terminal size in rows and columns or gets sensible defaults.

pub fn size_checked(&self) -> Option<(u16, u16)>

Returns the terminal size in rows and columns.

If the size cannot be reliably determined None is returned.

pub fn move_cursor_to(&self, x: usize, y: usize) -> Result<()>

Moves the cursor to x and y

pub fn move_cursor_up(&self, n: usize) -> Result<()>

Moves the cursor up n lines

pub fn move_cursor_down(&self, n: usize) -> Result<()>

Moves the cursor down n lines

pub fn move_cursor_left(&self, n: usize) -> Result<()>

Moves the cursor left n lines

pub fn move_cursor_right(&self, n: usize) -> Result<()>

Moves the cursor down n lines

pub fn clear_line(&self) -> Result<()>

Clears the current line.

The positions the cursor at the beginning of the line again.

pub fn clear_last_lines(&self, n: usize) -> Result<()>

Clear the last n lines.

This positions the cursor at the beginning of the first line that was cleared.

pub fn clear_screen(&self) -> Result<()>

Clears the entire screen.

pub fn clear_to_end_of_screen(&self) -> Result<()>

Clears the entire screen.

pub fn clear_chars(&self, n: usize) -> Result<()>

Clears the last char in the the current line.

pub fn set_title<T: Display>(&self, title: T)

Set the terminal title

pub fn show_cursor(&self) -> Result<()>

Makes cursor visible again

pub fn hide_cursor(&self) -> Result<()>

Hides cursor