Struct console::Term

pub struct Term { /* fields omitted */ }

Abstraction around a terminal.

A terminal can be cloned. If a buffer is used it's shared across all clones which means it largely acts as a handle.

Methods

impl Term

Important traits for Term
impl Write for Termimpl Read for Term

pub fn stdout() -> Term

Return a new unbuffered terminal

Important traits for Term
impl Write for Termimpl Read for Term

pub fn stderr() -> Term

Return a new unbuffered terminal to stderr

Important traits for Term
impl Write for Termimpl Read for Term

pub fn buffered_stdout() -> Term

Return a new buffered terminal

Important traits for Term
impl Write for Termimpl Read for Term

pub fn buffered_stderr() -> Term

Return a new buffered terminal to stderr

pub fn target(&self) -> TermTarget

Returns the targert

pub fn write_line(&self, s: &str) -> Result<()>

Writes a string to the terminal and adds a newline.

pub fn read_char(&self) -> Result<char>

Read a single character from the terminal

This does not echo the character and blocks until a single character is entered.

pub fn read_key(&self) -> Result<Key>

Read a single key form the terminal.

This does not echo anything. If the terminal is not user attended the return value will always be the unknown key.

pub fn read_line(&self) -> Result<String>

Read one line of input.

This does not include the trailing newline. If the terminal is not user attended the return value will always be an empty string.

pub fn read_secure_line(&self) -> Result<String>

Read securely a line of input.

This is similar to read_line but will not echo the output. This also switches the terminal into a different mode where not all characters might be accepted.

pub fn flush(&self) -> Result<()>

Flushes internal buffers.

This forces the contents of the internal buffer to be written to the terminal. This is unnecessary for unbuffered terminals which will automatically flush.

pub fn is_term(&self) -> bool

Checks if the terminal is indeed a terminal.

This is a shortcut for features().is_attended().

pub fn features(&self) -> TermFeatures

Checks for common terminal features.

pub fn want_emoji(&self) -> bool

Deprecated since 0.8.0
: Use features().wants_emoji() instead

Checks if this terminal wants emoji output.

pub fn size(&self) -> (u16, u16)

Returns the terminal size or gets sensible defaults.

pub fn size_checked(&self) -> Option<(u16, u16)>

Returns the terminal size in rows and columns.

If the size cannot be reliably determined None is returned.

pub fn move_cursor_up(&self, n: usize) -> Result<()>

Moves the cursor up n lines

pub fn move_cursor_down(&self, n: usize) -> Result<()>

Moves the cursor down n lines

pub fn clear_line(&self) -> Result<()>

Clears the current line.

The positions the cursor at the beginning of the line again.

pub fn clear_last_lines(&self, n: usize) -> Result<()>

Clear the last n lines.

This positions the cursor at the beginning of the first line that was cleared.

pub fn clear_screen(&self) -> Result<()>

Clears the entire screen.